+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Improvements
 
 - (app) [#2373](https://github.com/evmos/evmos/pull/2373) Remove ChainID constraint.
+- (rpc) [#2375](https://github.com/evmos/evmos/pull/2375) Add support for allow-insecure-unlock.
 
 ### State Machine Breaking